pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "python-whois>=0.9.5",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.2",
    "validators>=0.34.0",
    "whois>=1.20240129.2",
]
